+++ src/base/GatewayTrait.php
@@ -7,11 +7,8 @@
 
 namespace craft\commerce\base;
 
-use craft\commerce\elements\conditions\orders\DiscountOrderCondition;
-use craft\commerce\elements\conditions\orders\GatewayOrderCondition;
 use craft\elements\conditions\ElementConditionInterface;
 use craft\helpers\App;
-use craft\helpers\Json;
 use DateTime;
 
 /**
@@ -86,46 +83,5 @@ trait GatewayTrait
     public function getIsFrontendEnabled(bool $parse = true): bool|string|null
     {
         return $parse ? App::parseBooleanEnv($this->_isFrontendEnabled) : $this->_isFrontendEnabled;
-    }
-
-    /**
-     * Gets the order condition for this gateway
-     *
-     * @since 5.4.0
-     */
-    public function getOrderCondition(): ElementConditionInterface
-    {
-        /** @var DiscountOrderCondition $condition */
-        $condition = $this->_orderCondition ?? new GatewayOrderCondition();
-        $condition->mainTag = 'div';
-        $condition->name = 'orderCondition';
-
-        return $condition;
-    }
-
-    /**
-     * Sets the order condition for this gateway
-     *
-     * @since 5.4.0
-     */
-    public function setOrderCondition(ElementConditionInterface|string|array $condition): void
-    {
-        if (empty($condition)) {
-            $this->_orderCondition = null;
-            return;
-        }
-
-        if (is_string($condition)) {
-            $condition = Json::decodeIfJson($condition);
-        }
-
-        if (!$condition instanceof GatewayOrderCondition) {
-            $condition['class'] = GatewayOrderCondition::class;
-            /** @var GatewayOrderCondition $condition */
-            $condition = \Craft::$app->getConditions()->createCondition($condition);
-        }
-        $condition->forProjectConfig = true;
-
-        $this->_orderCondition = $condition;
     }
 }
